[Core] Remove inline entity reference methods
Closes #2.

These methods pass on the responsibility of parsing, tokenising and
expanding references in arbitrary strings to the manager with no context
of the formatting of the string. Understanding the myriad of character
escaping/quoting possibilities in these strings is far out-of-scope for
a manager.

Signed-off-by: Tom Cowland <tom@foundry.com>
Sign-off-executed-by: Tom Cowland <tom@foundry.com>
Approved-at: https://github.com/TheFoundryVisionmongers/OpenAssetIO/pull/85