[Core] Remove generic 'command' API
The command API was intended as an open-ended extension point for the
API, to safeguard future needs. Given its nature - in that, both a host
and a manager need to mutually agree on the definition of a command - does
it really have any place?

The whole point of this API is to define a common set of interactions.
It we need to rely on ad-hoc extensions via a blind interface, then hasn't
it failed?

The `CommandSpecification` infrastructure goes some way towards being
able to standardise the definition of a command (such that there is less
ambiguity), but if any given operation is that common, should it not be
a first-class part of the API anyway?

The only currently understood application would be to support `Ar 2.0`
methods that have no direct equivalent in OAIO. Given the potential
prevalence of that API, it feels like native interoperability, at the
conceptual level at least, would be a far superior way to achieve
interoperability.

Signed-off-by: Tom Cowland <tom@foundry.com>
Sign-off-executed-by: Tom Cowland <tom@foundry.com>
Approved-at: https://github.com/TheFoundryVisionmongers/OpenAssetIO/pull/85