[Docs] Add bare-bones Doxygen docs build
Migrates the FnAssetAPI Doxygen build across. We use Doxygen as it has a
proven history with mixed C++ and Python projects. When combined with
`doxypy` it allows us to write Doxygen formatted markup in docstrings
and have it turn into meaningful output in the resulting docs.

Doxygen was originally chosen over Sphinx as it includes a TOC for each
entry’s methods. We generally have a lot of documentation for each method,
and loosing the TOC makes it very challenging to navigate.

The main downside to this [current] setup, is that it blatantly exposes
the namespace hoisting we do in Python. This results in redundant nesting
for classes that are defined their own file. We’re putting up with this
right now, as the entire core API will be moved to `C++` soon, and this
issue will go away.

The HTML + CSS for this is by no means tidy, and was originally put
Together when IE9 was still a thing. It needs a refresh at some point,
but having a working build right now is the priority.

Part of #12.