[Docs] Hide namespace list
For some reason, highlighting in the sidebar TOC tree was broken inside
the namespace section. The nav would jump from the class list to the
namespace list, and highlight the wrong thing. Eg, clicking on `managerAPI`
would highlight `specifications`.

Taking the easy route for now, and just hiding the namespace list. At
least it makes the nav consistent. Practically, browsing the class list
vs the namespace list is no different.

Part of #34.