[Lint] Wrap docstrings to 72 characters
As specified in #41, we want to conform to PEP8 conventions where
possible. In particular, docstrings and documentation in general should
be limited to a 72 character line length.

Various automated tools attempted. Some like PyCharm and autopep8
refuse to modify docstrings. Also tried `style-doc`, which broke the
code (attempted to reformat `#  """`); and `docformatter`, which is
bugged to only work for the first summary line.

Eventually wrote a throwaway Python script to automatically line-wrap
our peculiar doxypy docstring notation.