[Lint] Add initial pyproject.toml
For #41 we want to establish good linter hygiene. Pragmatically, most
remaining linter violations will not be addressed right now (if only
because much of the codebase is due to be migrated to C++). So
currently the pylint config just contains basic configuration with
regexes for OAIO naming conventions.

Signed-off-by: David Feltell <david.feltell@foundry.com>
Sign-off-executed-by: Tom Cowland <tom@foundry.com>
Approved-at: https://github.com/TheFoundryVisionmongers/OpenAssetIO/pull/85