[Lint] Fix function signature for Doxypy parsing
For #41 an implicit requirement is not to make the existing Doxygen
rendered documentation any worse. The line wrapping performed by the
autoformat caused breaking changes to the html (which was already quite
broken because of the unsupported type hints).

So just remove type hints, which trip up doxypy when parsing to generate
documentation.

We will be removing all the type hints in upcoming work to fix #40
anyway, as well as leaving the door open for potential Python 2
compatibility.