[Lint] Fix broken Doxygen after docstring wrapping
Work so far on #41 caused a few scattered issues in the Doxygen rendered
documentation. So add some fixes based on a diff of doc html before and
after changes for #41 (so few pre-existing issues are addressed).

Doxygen is very sensitive to whitespace, including single newlines,
which results in some html rendering artifacts after line-wrapping.

The examples in `getRelatedEntities`/`getRelatedReferences` docstrings
are now cleaned up to fit on a single 72 char line. Additionally, they
are indented to ensure they render as code blocks rather than block
quotes.

For the `Host` constructor docstring, `@private` has no effect on the
rendered output, other than to add an empty `<pre>`, so remove it. As a
bonus, remove the type hint so that parameters are now rendered.

Signed-off-by: David Feltell <david.feltell@foundry.com>
Sign-off-executed-by: Tom Cowland <tom@foundry.com>
Approved-at: https://github.com/TheFoundryVisionmongers/OpenAssetIO/pull/85