[Core] Add `ValidatingMockManagerInterface`
For testing #43 we need some way of asserting that we've updated the
types that methods accept, since the current tests just assert that
whatever was given to `Manager` is passed through verbatim to
`ManagerInterface`, making TDD rather tricky.

So add a `ManagerInterface` implementation that is a hybrid of a mock
and a concrete class, such that asserts can be placed within the
methods.

Signed-off-by: David Feltell <david.feltell@foundry.com>