[Build] Extract docs deploy logic into script
In order to support CI builds (see #21), we need some more intelligent
handling of the "nothing to commit" scenario. Attempting to encapsulate
this in single-line `Makefile` commands becomes increasingly hard to
grok.

This splits out the deploy step into a separate script where the logic
is easier to define, and reason about.

Signed-off-by: Tom Cowland <tom@foundry.com>