[CI] Suppress summary from grep when checking for doc build errors
#104 introduced PR checks for the Doxygen docs build. Due to the unique
way in which problem matchers work, we require manual effort to actually
fail the job. We'd been using `! grep` for this purpose, and printing
its output as a summary of failures. However, some of the Doxygen
warnings are multi-line, and so the summary is less helpful.
This switches grep to quiet mode. The reader can always click-through
from the annotation to the correct line in the log anyway, which makes it
easier to see the remaining lines of the message until (if) we figure
out a multi-line matcher.

Signed-off-by: Tom Cowland <tom@foundry.com>