[Core] Add `kWantsThumbnail` flag to `managementPolicy`
Part of #86, adds a flag to `managementPolicy` that can be used to allow
a manager to request thumbnail generation as part of a publish.

`thumbnailSpecification` will be removed in a subsequent commit.

Updates documentation/examples to explain the new mechanism, which is
vastly superior to the old approach as it uses standardised API
interactions.

Signed-off-by: Tom Cowland <tom@foundry.com>