[Core] Rename `metadata` to `attributes`
Closes #3. Using the term "metadata" was fundamentally misleading. There
is nothing "meta" about the data held in an entity's associated key/value
pairs. It is considered part of the entity itself - as a companion to
the primary string.

Metadata also conjured up thoughts around timestamps, permissions etc.
when raised in discussions around the purpose of the API.

We considered 'data', but it was too ambiguous with the 'data' that may
be referenced by the primary string when it is an URL.

We considered 'properties', and enjoyed a journey down a philosophical
rabbit hole reading Wikipedia articles on the philosophy of attributes
and properties and ended up coming up none the wiser.

We settled on attributes, probably for no specific reason. The computing
version of the attributes wikipedia entry says this:

"In computing, an attribute is a specification that defines a property
of an object"

So attributes felt more fitting as within OAIO we desire to anchor
concrete meaning and usage to this data. I'm sure it'll be the wrong way
around which ever one we pick. 🙃

Signed-off-by: Tom Cowland <tom@foundry.com>