[Core] Make `ManagerInterface.identifier` static
It doesn't make much sense to allow the manager's identifier to vary.
It has to be consistent with the one reported by the ManagerPlugin.
Having this be an instance method makes having a single source of truth
a little ugly to no tangible benefit.

Leaving the host side, as it makes more sense to be an instance method,
as an application may have different modes served from the same binary
(eg: PLE).

Considered making this a read-only property, but keeping it as a method
for symmetry with `HostInterface` (which has to be a method in c++).

Signed-off-by: Tom Cowland <tom@foundry.com>