[Examples] Add skeleton for SampleAssetManager
Part of #96. Adds a skeleton SampleAssetManager plugin. This will form
the basis of an example manager implementation that can be used by
the test infrastructure, and by host developers to explore OAIO
functionality.

This has no explicit test as these will be provided by #95, which is
being developed in parallel.

Includes updated tests for the `PluginSystemManagerFactory` to ensure
it loads plugins provided in the source tree.

Signed-off-by: Tom Cowland <tom@foundry.com>