[Test] Add manager plugin test harness
Part of #95. Adds a skeleton test harness that can be used to validate
the implementation of any given `ManagerInterface` implementation.

We are building this in thin vertical slices to better facilitate #96.
Consequently, this version has minimal coverage of the bootstrap methods
`identifier`, `displayName` and `info`. This will be extended over time
to cover the full interface as we developer the example manager plugins.

TODO:
 - Finalise the public interface for this code when not used via the
   Command line. It should be considered private for now.
 - Add end-to-end tests that run the suite against the
   SampleAssetManager (SAM).

NB: Message updated during squash by Tom Cowland <tom@foundry.com>

Signed-off-by: David Feltell <david.feltell@foundry.com>