[Tests] Add manager plugin test harness
Closes #95. Adds a test harness that can be used to validate the
implementation of any given `ManagerInterface` implementation.

We are building this in thin vertical slices to better facilitate #96.
Consequently, this version has minimal coverage of the bootstrap methods
`identifier`, `displayName`, `info` and `managementPolicy`. This will be
extended over time to cover the full interface as we developer the
example manager plugins.

The `apiComplianceSuite` quickly highlighted that it's not easy to
define what should go in that suite. So much of what the a manager test
should cover includes results that are very specific to the _inputs_
to the methods, not the API contract itself.

We had two options:

 - One was to parameterise some more generalised tests for each method
   that allowed you to define n number of invocations with different
   arguments and their expected results.

 - Focus solely on the areas of the API that require specific knowledge
   of the semantics of the API contract and its invocation.

We settled on the latter, as the former is basically making a meta test
framework. This is kind of pointless and can be more fluently expressed
as standard test cases written by the manager author. None of the
resultant tests even address the main purpose of this suite, which is
"have I implemented my plugin right _by the API contract_?" - they're
telling you if you implemented it right by _your_ business logic.

Sticking to this smaller set of tests mean that the plugin authors don't
need to learn how to use a host to properly exercise their plugin.

To make it easier for a plugin developer to add new suites that do cover
local business logic, we also provide a public python API into the test
harness. This allows a developer to re-use the framework with their
own test suites with minimal effort.

Developed through #139, #179, #184, #186 and #217.

Co-authored-by: David Feltell <david.feltell@foundry.com>
Signed-off-by: Tom Cowland <tom@foundry.com>