[Tests] Suppress expected `pylint` errors from test cases
Part of #101. At present, our use of the when/then test naming
convention creates assorted `pylint` errors when our standard config
is used. At present, we don't have a good mechanism to support a
different `pylint` config for the `tests` subdirectory.

As such, our current approach is to silence some of these expected
errors in the test suite implementations themselves.

Signed-off-by: Tom Cowland <tom@foundry.com>