[Core] Ensure `Specification.data` contains un-set values
Previously, it would only return values that had been manually set,
and allowed access to underlying container.

Though this might be useful for efficient internal bridging between
languages, it breaks encapsulation, and makes determining the full
set of data for a spec more long winded than without it. We should
consider if we really even need this method.

Signed-off-by: Tom Cowland <tom@foundry.com>