[CI] Treat `C`, `I` and `R` category `pylint` warnings as fails
Part of #101. This re-enables C,I and R error categories for CI `pylint`
checks.

  "Your code has been rated at 10.00/10" (LOL)

We now only exclude `fixme` as we track TODOs though other means, and
allow them to exist beyond the lifetime of a PR. As such, they shouldn't
fail CI checks. This check should be considered a merge requirement
from this point forwards.

Signed-off-by: Tom Cowland <tom@foundry.com>