[ManagerValidator] Fix CLI exit code
Closes #160.

We use a self-contained, test-local `ManagerPlugin` for these tests
to avoid inadvertently coupling to the specifics of the
`SampleAssetManager`. It does mean there is a small overhead in
maintaining a minimal set of passing test fixtures. However, due to our
intended "minimal viable asset manager plugin" being one that says
it ignores all specifications, then this shouldn't be too onerous long
term.

Signed-off-by: Tom Cowland <tom@foundry.com>