[ManagerValidator] Migrate public API to `harness` module
A design goal of the manager test harness is to simplify the development
and testing of a manager plugin. The included default suite covers
compliance with the OpenAssetIO API contract, but is unable to test
implementation-specific logic. eg. a developer may wish to ensure that
their `managementPolicy` implementation is returning the expected
values.

To address this, we also provide a public python API into the test
harness. This allows a developer to re-use the framework with their
own test suites with minimal effort.

As part of #177, this commit moves the public API into a `harness`
module, and protects the remaining private parts of the implementation.

This formalises a stable, supported public API by making a clear
distinction between this and any private implementation classes.

Further documentation will be made available in a subsequent commit.

Signed-off-by: Tom Cowland <tom@foundry.com>