[Build] Fix setting default `CMAKE_INSTALL_PREFIX`
Part of #183.

As a convenience we set the default install location to
`dist` in the repo root, but only if the project isn't being included
as a CMake subproject, i.e. only if `PROJECT_IS_TOP_LEVEL`. However,
the minimum required CMake version didn't support `PROJECT_IS_TOP_LEVEL`
so that convenience config was ignored.

Signed-off-by: David Feltell <david.feltell@foundry.com>