[Build] Remove STATIC_DEFINE for export header
Part of #183. The generated export header content is different for
shared vs. static libs, meaning `STATIC_DEFINE` is redundant in our
(current) case. `STATIC_DEFINE` is useful when simultaneously building a
static and shared library, allowing the shared library's export header
to be reused for both.

If/when we support simultaneous shared and static builds we'll have to
carefully consider which libraries are using which export header with
which `#define`s.  In the meantime don't presume, just remove.

Signed-off-by: David Feltell <david.feltell@foundry.com>