[Docs] Add decision record for generic string type
Part of #193. Through various conversations the question of string
representation was surprisingly contentious. Several potential pitfalls
of naively using built-in types were alluded to, with regard to safety
and flexibility.

This DR records some further investigation and our thinking on how we
should approach string representation in C++, with consideration given
to language bindings in Python and C.

Signed-off-by: David Feltell <david.feltell@foundry.com>