[Build] Simplify Python install build target
For #183 convenience targets were added for creating a Python
environment with OpenAssetIO built and installed into it, useful for
running tests.

The original plan was to use `pip install --editable` for the pure
Python component, so that dev work wouldn't require a re-install.
However, this didn't work out since the editable install conflicted with
the CMake install of the binary Python extension, so the `--editable`
was removed.

This left some CMake cruft that was specific for `--editable` installs.
In particular, we used `add_custom_command` with an `OUTPUT` of the
`.egg-link` generated by the `--editable` install. The `.egg-link` is
not created for a non-editable install, so this is pointless - it just
had the effect of causing the `pip install` to always be executed (which
is what we want, but only coincidentally).

So tidy up to remove the now-pointless `add_custom_command` and roll it
into the `add_custom_target`.

Also remove `dev` from the target name, since that was originally
there to reflect the `--editable` nature.

Signed-off-by: David Feltell <david.feltell@foundry.com>