[CI] Update clang-format to use Google style
Part of #182.

In order to avoid debates on C++ coding style we wish to base our style
on one of the existing clang-format presets. After a vote within the
main dev team we decided on the Google style.

So reduce `.clang-format` to the be `BasedOnStyle: Google` with the only
override being the line length limit, plus hints on the language and
standard (updated from C++14 to C++17 as per current VFX Reference
Platform CY21/22)

Hence reformat the existing C++ code with this new style.

Signed-off-by: David Feltell <david.feltell@foundry.com>