[CI] Add Github action to check ClangFormat rules
Part of #182. Add a Github Action that runs `clang-format` on all
`.hpp`/`.h`/`.cpp`/`.c` files in `--dry-run` mode and error out
(`--Werror`) if any style problems are found.

Parsing the output to generate annotations for Github is left as a
future exercise (potentially one of the several third-party actions
could do this for us).

Signed-off-by: David Feltell <david.feltell@foundry.com>