[CI] Enable sanitizers when running tests in CI
Part of #182. A test run provides an ideal scenario for enabling
sanitizers that can check our code for difficult to detect bugs, such as
memory leaks.

Not all sanitizers can be used together, so enable only the address
sanitizer (ASan) and the undefined behaviour sanitizer, as these seem
the most important at this stage.

A little dance has to be done to get ASan working in libraries loaded
via a Python interpreter. So bake this into the `ctest` target that
wraps `pytest`.

The following lists are copied from the docs at
* https://github.com/google/sanitizers/wiki/AddressSanitizer and
* https://clang.llvm.org/docs/UndefinedBehaviorSanitizer.html
respectively.

The address sanitizer checks include
* Use after free (dangling pointer dereference).
* Heap buffer overflow.
* Stack buffer overflow.
* Global buffer overflow.
* Use after return.
* Use after scope.
* Initialization order bugs.
* Memory leaks.

The undefined behavior sanitizer checks include
* Array subscript out of bounds, where the bounds can be statically
  determined.
* Bitwise shifts that are out of bounds for their data type.
* Dereferencing misaligned or null pointers.
* Signed integer overflow.
* Conversion to, from, or between floating-point types which would
  overflow the destination.

Signed-off-by: David Feltell <david.feltell@foundry.com>