[Test] Add notes on the scope of `apiComplianceSuite`
Whilst working on #131, it quickly became apparent that it's not
easy to define what should go in that suite. So much of what the
a manager test should cover includes results that are very specific
to the _inputs_ to the methods, not the API contract itself.

We had two options:

 - One was to parameterise some more generalised tests for each method
   that allowed you to define n number of invocations with different
   arguments and their expected results.

 - Focus solely on the areas of the API that require specific knowledge
   of the semantics of the API contract and its invocation.

We settled on the latter, as the former is basically making a meta test
framework. This is kind of pointless and can be more fluently expressed
as standard test cases written by the manager author. None of the
resultant tests even address the main purpose of this suite, which is
"have I implemented my plugin right _by the API contract_?" - they're
telling you if you implemented it right by _your_ business logic.

Sticking to this smaller set of tests mean that the plugin authors don't
need to learn how to use a host to properly exercise their plugin.

Signed-off-by: Tom Cowland <tom@foundry.com>