[Test] Provide a suitable locale to `FixtureAugmentedTestCase`
Part of #131, the makes the test harness hosts follow best practice of
setting a suitably configured locale in all context's passed to API
calls. Doing it in the test loader saves the test author from having to
write a lot of boilerplate.

Ultimately, we should probably make `Context.locale` required and
enforce that is its not None somewhere.

Signed-off-by: Tom Cowland <tom@foundry.com>