[Tests] Improve `pluginsSystem` coverage using `SampleAssetManager`
The `SampleAssetManager` provided as part of #96 allows us to extend
the coverage of the plugin system.

NB: Long-term, we should perhaps adopt the strategy employed with other
tests - whereby there is a local stub plugin accompanying the test.
This would properly de-couple the tests of the plugin system from any
specifics of the supplied examples/resources.

Signed-off-by: Tom Cowland <tom@foundry.com>