[CI] Add Clang-Tidy and Cpplint as part of build
Part of #182. Add options that, if enabled, execute `clang-tidy` and
`cpplint` as part of the build process. This is handy since it can be
used in both CI and on dev machines, so that devs can be confident that
CI won't fail when they push.

Clang-Tidy is a static analysis checker that uses LLVM magic to assert
on various style issues and bug-prone code.

Cpplint is a much simpler linter written in Python, based on Google's
style guide. One specific check that sold it for us is to assert that
all files have a copyright header, which Clang-Tidy has no opinion on.

For `clang-tidy`, CMake's built-in integration is pretty good, so use
that, configured to fail the build if any checks fail.

The corresponding CMake integration of `cpplint` is less impressive. It
won't check headers and cannot be configured to fail the build. So for
this case roll our own target to execute the `cpplint` on the
command line, which all other targets should depend on.

Hence fix a couple of issues discovered by these linters in the C++
sources.

Signed-off-by: David Feltell <david.feltell@foundry.com>