[CI] Integrate C++ linters into CI
Part of #182. Add a CI job to execute `clang-tidy` and `cpplint`, using
a similar structure to the unit test job.

Hence update the bootstrap script to ensure these tools are available
in CI (and Vagrant VMs).

Ensure the same Github cache key is used for the Conan package cache
across both test and linter jobs (`runner.os` was "Linux" vs.
`matrix.os` which is "ubuntu-20.04"), so they can reuse each other's
cache.

Disable `OPENASSETIO_WARNINGS_AS_ERRORS` on unit test execution, which
fails the build if any compiler warnings are flagged, and instead add
this to the linter job. This means tests will still run even if there
are compiler warnings, reducing CI churn.

Also simplify the unit test job to remove pointless `-Wdev`, which is
for developer-specific CMake warnings, which we'll never see in CI (but
should enable for local dev).

Signed-off-by: David Feltell <david.feltell@foundry.com>