[Core] Remove optional entity ref from `managementPolicy`
A partial precursor to #115, removes the optional `entityRef` from
`managementPolicy`. It seemed like a good idea at the time, but in
reality, just opens up ambiguity, and only half solves the problem.

#9 covers the fact that we need a broader mechanism for working out
if operation X can be done in the context of Y. One of these things
could be to register a new entity to a specific target entity. A more
flexible mechanism that covers this and the topics in #9 would be
far better than shoehorning the check into `managementPolicy` - which
has no mechanism for communicating why.

With the optional reference, it was not clear if a host should always
call this prior to `preflight`/`register` or not.
Technically, it should! When we do resolve #9, that would then mean two
methods potentially needed to be called before a registration/preflight,
which would be a very poor design indeed.

By making `managementPolicy` entity invariant, it makes its use much
clearer as to when it should be called by a host, and much easier for
a host to respect its response.

It can now be simply defined as a one-off check for each context access
pattern in relation to any given specification to enable or disable
host-side behaviours.

Signed-off-by: Tom Cowland <tom@foundry.com>