[Tests] Add support for shared fixtures in `test.manager`
Whilst extending the `apiComplicanceSuite` to cover more of the
`ManagerInterface` surface are as part of #215, it quickly became
apparent that requiring fixtures to be declared at the function level,
was not even remotely DRY and quickly became cumbersome, and hard to
read.

This adds support for `"shared"` fixture entries that are inherited by
classes and then in turn, functions.

An additional minor change is that within a `FeatureAugmentedTestCase`,
`self._fixtures` is always set to a `dict` instead of potentially None.
This shouldn't cause any problems detecting missing fixture values as a
key lookup will still fail if it is missing.

Signed-off-by: Tom Cowland <tom@foundry.com>