[Tests] Add ability to load test suite module from arbitrary files
When writing tests for a new manager, as part of #215, it highlighted
that a significant hurdle in easily re-using the
`openassetio.test.manager` harness is that in order to `import` your
test suite, it needs to be on `PYTHONPATH`. Because the system under
test is invariably an OpenAssetIO `ManagerPlugin`, the code itself
doesn't actually need to be on `PYTHONPATH` itself - the plugin system
takes care of all that for you. Consequently, the test suite you want to
import is not going to be on python path either (especially if it is an
out-of-source suite).

This commit adds a simple utility function that can load a module from
a file, so that custom test harness code can easily import a test suite
from a file local to the rest of the test code.

Signed-off-by: Tom Cowland <tom@foundry.com>