[Tests] Add fixture validation methods to `FixtureAugmentedTestCase`
Through the development of tests for #215, we found ourselves writing
a lot of additional boilerplate to either validate that fixtures had
been provided, or to skip a test if they weren't. @feltech had the
good idea to generalise this mechanism.

This commit introduces two (four) new methods to the
`FixtureAugmentedTestCase`, that allows expected fixture values to be
retrieved whist causing the test to be skipped or failed if any of the
requested fixtures are missing.

This is provided in two forms:

   `collectRequiredFixtures`: Stores the fixture values in instance
      variables, useful in setUp where the same fixtures are used in
      multiple individual tests, as it removes the need to duplicate
      calls in each individual test.

   `requireFixtures`: Returns fixture values.This can be used in-line
      where normal dict lookups where being used, or in programatic
      cases.

We err on the side of being "over helpful" with the boilerplate methods
for single fixture lookup. This is a common case, and the extra list
wrapping and subscripting needed to extract single values doesn't help
with readability and can be error prone when developing new test cases.

As this is a public API to be used by plugin authors, it felt like they
were worth while.

Signed-off-by: Tom Cowland <tom@foundry.com>