[Examples] Fix path parsing in `bal:///` urls
An inadvertent side effect of the fact that:

  urllib.parse.urlparse("aString")

Results in:

  ParseResult(scheme='', netloc='', path='aString', ...)

This meant that an error in the BAL fixtures that missed off `bal:///`
went unnoticed. Once this was rectified, the tests then correctly
highlighted that the BAL code wan't removing the leading / from the path
parsed from the BAL entity references.

Part of #215

Signed-off-by: Tom Cowland <tom@foundry.com>