[Core] Add specification/trait C++ implementation
For #250 we require the C++ infrastructure to support specifications
following the new design of #238 et al. This commit introduces an
initial draft of specifications as dictionary-like data structures and
traits as views on a specification, coded in C++ with Python bindings,
and tested exclusively through the Python bindings.

The `Specification` class uses the pimpl idiom to hide the underlying
implementation. This class centralises the data that needs transporting
in order to reproduce a specification, both it's properties and
supported traits.

`Specification`'s constructor takes a list of traits that it supports.
The underlying dict of a specification is initialised at construction
with an entry for each supported trait.

Initial performance tests using #270 and reported in #253 lead us to
believe the performance of this design is unlikely to be a serious
bottleneck.

Properties are accessed and mutated on the `Specification` via a
`std::variant` of the allowed value types. For typed access (the assumed
usual case), we expect trait views to be used rather than directly
accessing the specification.

`TraitBase` provides the common interface for a trait view and is not
expected to be used independently, i.e. for polymorphism or in language
bindings. Hence it is a non-polymorphic CRTP class with the aim of
reducing the boilerplate required to write trait classes in C++.

`BlobTrait` is included as an example concrete trait view, and may be
removed in the future.

The types of properties available to store within specifications are
aliased for readability and maintainability in a `trait::property`
namespace.

The `TraitPropertyStatus` enum is used to differentiate property access
error types ("unset" vs."wrong value type") without resorting to
exceptions. This allows less defensive code by hosts/managers if they
are happy to silently ignore failures. This necessitates the use of an
out-parameter pattern, so that the return value can be used for a
`TraitPropertyStatus`.

Currently, the exception to the use of `TraitPropertyStatus` is when
querying/mutating properties on a specification when the trait is not
supported (doesn't already exist in the underlying dictionary).  In this
case a `std::out_of_range` exception will be thrown (via
`std::unordered_map::at`).

Python binding sources are suffixed with "Binding". This makes it easier
to tell at a glance (e.g. in a global file search) the responsibility of
the file.

Python bindings use a `std::shared_ptr` holder type, since we expect
ownership to be shared between C++ and Python, and wish to avoid
expensive copies. This is probably beneficial for `Specification`, but
may not be necessary for traits - a subject of further design work.

For Python traits (only `BlobTrait` at this time), add a `raiseOnError`
keyword argument, which if `True` will raise the appropriate exception
for `TraitPropertyStatus`, and if `False` will cause `None` to be
returned for any error status. Hence `TraitPropertyStatus` is not bound
explicitly.

Add argument names to improve pybind auto-generated docstrings.
Otherwise leave Python docstrings empty, for now. Users must refer to
the C++ documentation and higher-level workflow example docs.  E.g.
```
>>> help(specification.Specification.getTraitProperty)
Help on instancemethod in module openassetio._openassetio.specification:
getTraitProperty(...)
    getTraitProperty(self: openassetio._openassetio.specification.Specification, traitId: str, key: str) -> Optional[Union[bool, int, float, str]]
```

Ensure pylint doesn't complain that it can't import our C++ module by
suppressing `import-error`, for now. This will be addressed in future
PRs.

Signed-off-by: David Feltell <david.feltell@foundry.com>