[Core] Remove the host document API
We still fundamentally believe that the ability for a manager to query
the current host for information about entity usage and its documents
is critical. However, the current API was not well thought out (as it
assumed a 'current' document, which may be a little tenuous).

This commit closes #6 by removing the current API, so that we can add
a more fit-for-purpose API in over time as we have more concrete use
cases from manager authors.

Signed-off-by: Tom Cowland <tom@foundry.com>