[Core] Improve context thawing API
Changes TransactionCoordinator.thawManagerState to be the reverse of
freezeManagerState.

The main motivation is to remove a redundant state creation when the
host creates a context, then immediately thaws a serialized state back
into it, overwriting the one just made by the manager.

This also makes the API more closely illustrate the conceptual model. In
that, the a Context is the host’s mechanism for tying together
associated API calls. The host manages the lifetime of the context such
that it is re-used for all conceptually related calls. The freeze/thaw
mechanism allows for a context to be serialized, distributed and
restored to associate API calls across time/processes.

The managerInterfaceState token is the way the API middleware bridges
the host side context-lifetime pattern into the stateless, re-entrant
world of the ManagerInterface. The host doesn’t know about the specifics
of this. All it knows is that a context can be frozen and thawed.

The old version of this API made it unclear that thawing a state token
into an existing context, detached that context from the “logical group”
that it was part of before. If you made a context, used it in an API
call then thawed into it, and used it again. The manager would not see
it as the same context, even though from the hosts perspective, it is
the same context based on the policy of lifetime management.

The updated API hopefully makes it clearer that thawing returns some
other context that is logically the same as the supplied token, not
whatever was passed into the method.

Closes #79.

Signed-off-by: Tom Cowland <tom@foundry.com>