[Core] Add cpp-cpp path test harness executable
This adds a starting point for a light-weight test harness that can be
used to exercise the cpp - cpp path to measure performance.
The initial goal of this harness was assessing the relative performance
of the trait-centric resolution mechanism compared to the previous
primary-string mechanism.

The perftest executable measures wall clock time for minimal
implementations of blob trait resolution vs resolveEntityReference
for 100,000 unique entity references with a "database" of 1 million
asset entries. The order of resolutions is randomized every run.

Building the perftest executable is disabled by default,
enable it in the project with the BUILD_PERFTEST CMake option.

Closes #260

Signed-off-by: Raihan Kibria <raihan.kibria@foundry.com>