[Core] Bump pybind11 version to 2.8.1
Use the same version as is baked into the CY2022 AWSF Docker image.
See: https://github.com/AcademySoftwareFoundation/aswf-docker/tree/master/ci-base

Amongst other advantages, this version adds a handy
`pybind11::attribute_error` exception that can be thrown from C++,
appearing in Python as `AttributeError`.

The initial motivation for this PR was to allow use of `attribute_error`
in #255, but aligning with ASWF is a compelling reason in its own right.

Signed-off-by: David Feltell <david.feltell@foundry.com>