[Core] Pass trait IDs to `managementPolicy`
Update the signature for `magementPolicy` to take a set of trait IDs
rather than a `Specification`. This removes the ambiguity around
whether or not the method should consider the properties of the traits
vs their identifiers.

managementPolicy is meant to be a high-level “are you interested” query
used to enable/disable host functionality at a board level rather than
in any entity-specific way. For example, a manager that is not
Interested in managing images should be left out of 90% of operations
of a compositor…

Closes #280.

Signed-off-by: Tom Cowland <tom@foundry.com>